requirements.txt
customtkinter
psutil
yara-python
requests
watchdog
pywin32
pefile
scikit-learn